moved embed into the tools directory and renamed utils to fonts
