+++ src/tools/embed.c
