+++ src/tools/wrenembed.c
